refactor: :hammer: reorganizada estrutura de pastas para melhor modularidade
- Criada uma pasta separada para o componente Input dentro de src/components/Input.
- Movido o arquivo Input.jsx e Form.css para a nova pasta src/components/Input.
- Atualizado o caminho de importação do componente Input no arquivo Form.jsx.
- Estrutura de pastas ajustada para melhorar a organização e modularidade do projeto.
- Testado localmente para garantir que os ajustes funcionem corretamente.
